Massive refactor. Not tested on CUDA, so there are surely bugs. Creates new directories: arguments (for runtime arguments for each script), data, model, conversions (contains all export/compile/onnx/etc transformation code), and utils.
