+++ quantization/QAT.py
@@ -1,15 +1,39 @@
+import argparse
 import logging
 import time
+from typing import Callable
 
+import torch
 import torch.optim as optim
 import torch.quantization as tq
 from torch.optim.lr_scheduler import StepLR
 
 
-def QAT(train, test, args, model, device, train_loader, test_loader):
+def QAT(
+    train: Callable,
+    test: Callable,
+    args: argparse.Namespace,
+    model: torch.nn.Module,
+    device: torch.device,
+    train_loader: torch.utils.data.DataLoader,
+    test_loader: torch.utils.data.DataLoader,
+):
     """
     Perform QAT on the model using the given dataloaders, and train adn test
     functions.
+    This works in place on the model.
+
+    Inputs:
+    - train (function): The training function.
+    - test (function): The testing function.
+    - args (argparse.Namespace): The command line arguments.
+    - model (torch.nn.Module): The model to quantize.
+    - device (torch.device): The device to run the model on.
+    - train_loader (DataLoader): The DataLoader to get a sample of data from.
+    - test_loader (DataLoader): The DataLoader to get a sample of data from.
+
+    Returns:
+    None
     """
 
     # Reset the optimizer with included quantization parameters
@@ -26,6 +50,7 @@ def QAT(train, test, args, model, device, train_loader, test_loader):
         if hasattr(module, "observer_enabled") or hasattr(module, "static_enabled"):
             module.disable_observer()
 
+    # QAT training lopp
     start_time = time.time()
     for epoch in range(1, args.epochs + 1):
         train(args, model, device, train_loader, optimizer, epoch)
