+++ utils/setup_logging.py
@@ -1,0 +1,23 @@
+import logging
+
+
+def setup_logging(log_file: str = None) -> None:
+    """
+    Sets up logging to print to console and optionally to a file.
+
+    Args:
+        log_file (str): Path to a log file. If None, logs will not be saved to a file.
+    """
+    log_format = "%(asctime)s - %(levelname)s - %(message)s"
+    date_format = "%Y-%m-%d %H:%M:%S"
+    log_level = logging.INFO
+
+    logging.basicConfig(
+        level=log_level,
+        format=log_format,
+        datefmt=date_format,
+        handlers=[
+            logging.StreamHandler(),
+            logging.FileHandler(log_file) if log_file else logging.NullHandler(),
+        ],
+    )
